Finish hook movement

+++ Assets/Scripts/Hook/HookConfig.cs
@@ -7,5 +7,5 @@ public class HookConfig : ScriptableObject
 	[field: Header("Movement")]
 	[field: SerializeField, Min(0f)] public float MoveSpeed { get; private set; } = 3f;
 	[field: SerializeField, Min(0f)] public float TileSize { get; private set; } = 1f;
-	[field: SerializeField, Min(0f)] public float MaxDistance { get; private set; } = 10f;
+	[field: SerializeField, Min(0f)] public int MaxDistance { get; private set; } = 10;
 }
